Fixed delete for all

+++ api/authors/delete.php
@@ -1,32 +1,43 @@
 <?php
-  // Headers
-  header('Access-Control-Allow-Origin: *');
-  header('Content-Type: application/json');
-  header('Access-Control-Allow-Methods: DELETE');
-  header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization,X-Requested-With');
+// Headers
+header('Access-Control-Allow-Origin: *');
+header('Content-Type: application/json');
+header('Access-Control-Allow-Methods: DELETE');
+header('Access-Control-Allow-Headers: Access-Control-Allow-Headers, Content-Type, Access-Control-Allow-Methods, Authorization, X-Requested-With');
 
-  include_once '../../config/Database.php';
-  include_once '../../models/Author.php';
-  // Instantiate DB & connect
-  $database = new Database();
-  $db = $database->connect();
+include_once '../../config/Database.php';
+include_once '../../models/Author.php';
 
-  // Instantiate blog post object
-  $author = new Author($db);
+// Instantiate DB & connect
+$database = new Database();
+$db = $database->connect();
 
-  // Get raw posted data
-  $data = json_decode(file_get_contents("php://input"));
+// Instantiate Author object
+$author = new Author($db);
 
-  // Set ID to UPDATE
-  $author->id = $data->id;
+// Get raw posted data
+$data = json_decode(file_get_contents("php://input"));
 
-  // Delete post
-  if($author->delete()) {
-    echo json_encode(
-      array('message' => 'Author deleted')
-    );
-  } else {
-    echo json_encode(
-      array('message' => 'Author not deleted')
-    );
-  }
+// Check if ID is provided
+if (!isset($data->id) || intval($data->id) <= 0) {
+    http_response_code(400); // Bad request
+    echo json_encode(['message' => 'Missing or invalid ID']);
+    exit;
+}
+
+// Set ID for deletion
+$author->id = intval($data->id);
+
+// Attempt to delete author
+if ($author->delete()) {
+    http_response_code(200); // OK
+    echo json_encode([
+        'id' => $author->id,
+        'message' => 'Author deleted'
+    ]);
+} else {
+    http_response_code(404); // Not found
+    echo json_encode([
+        'message' => 'No author found with the specified ID'
+    ]);
+}
